feat: add hook for adapting setup configurations
...that is executed before the configuration is used in any way (e.g. to
create tasks). That allows adapting the configuration not only before
task execution (which is possible with `afterEvaluate`) but before task
configuration.

An example use case are adapting configuration settings that influence
if a certain image should be built or not. For that the configuration
needs to be adapted before the build tasks are created.

ING-2940