feat: support explicit access to local/relative configuration
As resolving configuration variables via the configuration root now
takes precedence over resolving variables locally, the underscore was
added as a special key to access the local configuration.

This works as long as the underscore key is not used directly in the
root or local configuration. In that case the direct access via root or
local configuration takes precedence.

ING-2940